+++ lib/over_react_test.dart
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 export 'src/over_react_test/common_component_util.dart';
+export 'src/over_react_test/console_log_utils.dart';
 export 'src/over_react_test/custom_matchers.dart';
 export 'src/over_react_test/dom_util.dart';
 export 'src/over_react_test/jacket.dart';
 export 'src/over_react_test/js_component.dart';
-export 'src/over_react_test/prop_type_util.dart';
 export 'src/over_react_test/react_util.dart';
 export 'src/over_react_test/validation_util.dart';
 export 'src/over_react_test/wrapper_component.dart';
