+++ test/over_react_test.dart
@@ -20,10 +20,10 @@ import 'package:react/react_client.dart';
 import 'package:test/test.dart';
 
 import 'over_react_test/common_component_util_test.dart' as common_component_util_test;
+import 'over_react_test/console_log_utils_test.dart' as prop_type_util_test;
 import 'over_react_test/custom_matchers_test.dart' as custom_matchers_test;
 import 'over_react_test/dom_util_test.dart' as test_util_dom_util_test;
 import 'over_react_test/jacket_test.dart' as jacket_test;
-import 'over_react_test/prop_type_util_test.dart' as prop_type_util_test;
 import 'over_react_test/react_util_test.dart' as react_util_test;
 import 'over_react_test/validation_util_test.dart' as validation_util_test;
